Move device specific things to devices sub folder
