+++ hwilib/devices/digitalbitbox.py
@@ -10,9 +10,9 @@ import os
 import binascii
 import logging
 
-from .hwwclient import HardwareWalletClient
-from .serializations import CTransaction, PSBT, hash256, hash160, ser_sig_der, ser_sig_compact, ser_compact_size
-from .base58 import get_xpub_fingerprint, decode, to_address, xpub_main_2_test
+from ..hwwclient import HardwareWalletClient
+from ..serializations import CTransaction, PSBT, hash256, hash160, ser_sig_der, ser_sig_compact, ser_compact_size
+from ..base58 import get_xpub_fingerprint, decode, to_address, xpub_main_2_test
 
 applen = 225280 # flash size minus bootloader length
 chunksize = 8*512
