+++ hwilib/devices/keepkey.py
@@ -1,13 +1,13 @@
 # KeepKey interaction script
 
-from .hwwclient import HardwareWalletClient
+from ..hwwclient import HardwareWalletClient
 from keepkeylib.transport_hid import HidTransport
 from keepkeylib.client import KeepKeyClient as KeepKey
 from keepkeylib import tools
 from keepkeylib import messages_pb2, types_pb2 as proto
 from keepkeylib.tx_api import TxApi
-from .base58 import get_xpub_fingerprint, decode, to_address, xpub_main_2_test
-from .serializations import ser_uint256, uint256_from_str
+from ..base58 import get_xpub_fingerprint, decode, to_address, xpub_main_2_test
+from ..serializations import ser_uint256, uint256_from_str
 
 import binascii
 import json
@@ -175,7 +175,7 @@ class KeepKeyClient(HardwareWalletClient):
 
     # Must return a base64 encoded string with the signed message
     # The message can be any string
-    def sign_message(self, message):
+    def sign_message(self, message, keypath):
         raise NotImplementedError('The HardwareWalletClient base class does not '
             'implement this method')
